+++ CHANGELOG.md
@@ -1,7 +1,15 @@
-## 0.2.0 - [01-14-2025] - davidlramirez95
+## 0.3.0 - [2025-01-23] - davidlramirez95
+### Added
+- Added the resource creation of the CORS module with different resource name
+- Added `.gitignore`
+
+### Removed
+- Removed `squidfunk/terraform-aws-api-gateway-enable-cors` module
+
+## 0.2.0 - [2025-01-14] - davidlramirez95
 ### Updated
 - Updated `x-blueflag-service-name` to `X-Blueflag-Service-Name`
 
-## 0.1.0 - [01-14-2025] - davidlramirez95
+## 0.1.0 - [2025-01-14] - davidlramirez95
 ### Added
 - API Gateway cors module